requirements.txt
uvicorn
fastapi
pyyaml

